feat(e2e): contractor domain — scenarios + spec rewrites
