+++ e2e/scenarios/contractor/contractor-onboarding.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "../../schema/scenario.schema.json",
+  "name": "Contractor onboarding flow",
+  "description": "Company set up for contractor onboarding.",
+  "domain": "contractor",
+  "baseDemo": "react_sdk_demo_contractor_onboarding",
+  "decorations": {},
+  "expectedContext": ["companyId"]
+}
